Add functionality to create new points of interest
Added CreatePointOfInterest method in PointsOfInterestController enabling us to create new points of interest for a given city. It accepts a PointOfInterestForCreationDto object, generates the new PointOfInterest and adds it to the city's PointsOfInterest collection. Also, cleaned up GetCities method in CitiesController and provided Name attribute to GetPointOfInterest action. A new README.md file was created providing details about the project, and a new model PointOfInterestForCreationDto was introduced.